functionsV2/dev_v2_saveEmployeeSkill/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
functions-framework==3.5.0

